+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project changes will be shown here.
 
 ## [unreleased]
 - Issue 56 - CNC004
+- Issue 129 - CNC021
 
 ## [1.0.1] - 2017-12-14
 - Copy Username in password List.