+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This project changes will be shown here.
 
 ## [unreleased]
+- Issue 125 - CNC017 - New SR from SO to Imt queue
 - Issue 56  - CNC004
 - Issue 149 - Create a Daily Report for P5 with sales orders to be sent to nosales@